Delete old inventory management and rename the new ones.
