+++ enderio-machines/src/main/java/com/enderio/machines/common/blockentity/StirlingGeneratorBlockEntity.java
@@ -6,9 +6,8 @@ import com.enderio.base.common.blockentity.sync.FloatDataSlot;
 import com.enderio.base.common.blockentity.sync.SyncMode;
 import com.enderio.machines.common.MachineTier;
 import com.enderio.machines.common.blockentity.base.PowerGeneratingMachineEntity;
-import com.enderio.machines.common.io.item.MachineInventoryLayout;
 import com.enderio.machines.common.init.MachineCapacitorKeys;
-import com.enderio.machines.common.io.item.NInventoryLayout;
+import com.enderio.machines.common.io.item.MachineInventoryLayout;
 import com.enderio.machines.common.menu.StirlingGeneratorMenu;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Inventory;
@@ -72,8 +71,8 @@ public abstract class StirlingGeneratorBlockEntity extends PowerGeneratingMachin
     }
 
     @Override
-    public NInventoryLayout getInventoryLayout() {
-        return NInventoryLayout.builder(getTier() != MachineTier.Simple)
+    public MachineInventoryLayout getInventoryLayout() {
+        return MachineInventoryLayout.builder(getTier() != MachineTier.Simple)
             .inputSlot((slot, stack) -> ForgeHooks.getBurnTime(stack, RecipeType.SMELTING) > 0)
             .build();
     }
