+++ enderio-machines/src/main/java/com/enderio/machines/common/blockentity/base/PoweredMachineEntity.java
@@ -11,7 +11,6 @@ import com.enderio.machines.common.MachineTier;
 import com.enderio.machines.common.blockentity.sync.CapacitorDataSlot;
 import com.enderio.machines.common.io.energy.MachineEnergyStorage;
 import com.enderio.machines.common.io.item.MachineInventoryLayout;
-import com.enderio.machines.common.io.item.NInventoryLayout;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
@@ -175,7 +174,7 @@ public abstract class PoweredMachineEntity extends MachineBlockEntity {
      * Whether the machine requires a capacitor to operate.
      */
     public boolean requiresCapacitor() {
-        NInventoryLayout layout = getInventoryLayout();
+        MachineInventoryLayout layout = getInventoryLayout();
         if (layout == null)
             return false;
         return layout.supportsCapacitor();
@@ -191,7 +190,7 @@ public abstract class PoweredMachineEntity extends MachineBlockEntity {
     }
 
     public ItemStack getCapacitorItem() {
-        NInventoryLayout layout = getInventoryLayout();
+        MachineInventoryLayout layout = getInventoryLayout();
         if (layout == null)
             return ItemStack.EMPTY;
         return getInventory().getStackInSlot(layout.getCapacitorSlot());
@@ -222,7 +221,7 @@ public abstract class PoweredMachineEntity extends MachineBlockEntity {
             return;
         }
 
-        NInventoryLayout layout = getInventoryLayout();
+        MachineInventoryLayout layout = getInventoryLayout();
         if (requiresCapacitor() && layout != null) {
             cachedCapacitorData = CapacitorUtil.getCapacitorData(getCapacitorItem()).orElse(DefaultCapacitorData.NONE);
         } else {
