+++ enderio-machines/src/main/java/com/enderio/machines/common/menu/MachineSlot.java
@@ -1,23 +1,20 @@
 package com.enderio.machines.common.menu;
 
 import com.enderio.machines.common.io.item.MachineInventory;
-import com.enderio.machines.common.io.item.NMachineInventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.items.SlotItemHandler;
 import org.jetbrains.annotations.NotNull;
 
-import javax.annotation.Nonnull;
-
 public class MachineSlot extends SlotItemHandler {
 
-    public MachineSlot(NMachineInventory itemHandler, int index, int xPosition, int yPosition) {
+    public MachineSlot(MachineInventory itemHandler, int index, int xPosition, int yPosition) {
         super(itemHandler, index, xPosition, yPosition);
     }
 
     @Override
-    public NMachineInventory getItemHandler() {
-        return (NMachineInventory) super.getItemHandler();
+    public MachineInventory getItemHandler() {
+        return (MachineInventory) super.getItemHandler();
     }
 
     @Override
